requirements.txt
python-dotenv==1.1.0
langchain_astradb==0.6.0
langchain==0.3.23
pypdf==5.4.0
google==3.0.0
google-generativeai==0.8.5
langdetect==1.0.9
google-api-python-client==2.167.0
sumy==0.11.0